# docs/swagger.yaml
swagger: "2.0"
info:
  title: "API Raízes do Nordeste"
  description: "Sistema Multicanal de Gestão de Pedidos — canais APP, TOTEM, BALCAO, PICKUP e WEB"
  version: "1.0.0"
basePath: "/"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Cole aqui: Bearer SEU_TOKEN"
...
paths:
  /auth/register:
    post:
      tags: ["Auth"]
      summary: "Cadastrar novo usuário"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            required: ["nome", "email", "senha", "consentimento_lgpd"]
            properties:
              nome:
                type: "string"
                example: "Maria Silva"
              email:
                type: "string"
                example: "maria@email.com"
              senha:
                type: "string"
                example: "Senha@123"
              perfil:
                type: "string"
                enum: ["ADMIN", "GERENTE", "CLIENTE", "COZINHA", "ATENDENTE"]
                example: "CLIENTE"
              consentimento_lgpd:
                type: "boolean"
                example: true
      responses:
        201:
          description: "Usuário criado"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                example: 1
              nome:
                type: "string"
                example: "Maria Silva"
              email:
                type: "string"
                example: "maria@email.com"
              perfil:
                type: "string"
                example: "CLIENTE"
        409:
          description: "E-mail já cadastrado"
        422:
          description: "Dados inválidos"
  /auth/login:
    post:
      tags: ["Auth"]
      summary: "Fazer login e obter token JWT"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            required: ["email", "senha"]
            properties:
              email:
                type: "string"
                example: "admin@raizes.com"
              senha:
                type: "string"
                example: "Admin@2026"
      responses:
        200:
          description: "Login realizado com sucesso"
          schema:
            type: "object"
            properties:
              accessToken:
                type: "string"
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              tokenType:
                type: "string"
                example: "Bearer"
              expiresIn:
                type: "integer"
                example: 3600
              usuario:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  nome:
                    type: "string"
                    example: "Admin Raízes"
                  perfil:
                    type: "string"
                    example: "ADMIN"
        401:
          description: "Credenciais inválidas"
  /auth/me:
    get:
      tags: ["Auth"]
      summary: "Ver meu perfil"
      security:
        - Bearer: []
      responses:
        200:
          description: "Dados do usuário autenticado"
        401:
          description: "Não autenticado"
...